+++ src/plugin-marklogic/main/uk/co/reecedunn/intellij/plugin/marklogic/xray/runner/XRayTestRunState.kt
@@ -15,14 +15,70 @@
  */
 package uk.co.reecedunn.intellij.plugin.marklogic.xray.runner
 
+import com.intellij.execution.DefaultExecutionResult
 import com.intellij.execution.ExecutionResult
 import com.intellij.execution.Executor
-import com.intellij.execution.configurations.RunProfileState
+import com.intellij.execution.executors.DefaultDebugExecutor
+import com.intellij.execution.executors.DefaultRunExecutor
+import com.intellij.execution.process.ProcessHandler
 import com.intellij.execution.runners.ExecutionEnvironment
 import com.intellij.execution.runners.ProgramRunner
+import com.intellij.execution.ui.ConsoleView
+import uk.co.reecedunn.intellij.plugin.marklogic.xray.configuration.XRayTestConfiguration
+import uk.co.reecedunn.intellij.plugin.marklogic.xray.test.XRayTestService
+import uk.co.reecedunn.intellij.plugin.processor.intellij.execution.configurations.RunProfileStateEx
+import uk.co.reecedunn.intellij.plugin.processor.intellij.execution.process.RunnableQueryProcessHandler
+import uk.co.reecedunn.intellij.plugin.processor.intellij.execution.ui.QueryConsoleView
+import uk.co.reecedunn.intellij.plugin.processor.intellij.execution.ui.results.QueryTextConsoleView
+import uk.co.reecedunn.intellij.plugin.processor.query.Query
+import uk.co.reecedunn.intellij.plugin.processor.query.RunnableQuery
+import uk.co.reecedunn.intellij.plugin.processor.query.RunnableQueryProvider
+import uk.co.reecedunn.intellij.plugin.xquery.intellij.lang.XQuery
 
-class XRayTestRunState(private val environment: ExecutionEnvironment) : RunProfileState {
-    override fun execute(executor: Executor?, runner: ProgramRunner<*>): ExecutionResult? {
-        return null
+class XRayTestRunState(private val environment: ExecutionEnvironment) : RunProfileStateEx {
+    override fun execute(executor: Executor?, runner: ProgramRunner<*>): ExecutionResult {
+        val processHandler = createProcessHandler(createQuery())
+        val console = createConsole(executor!!)
+        console.attachToProcess(processHandler)
+        return DefaultExecutionResult(console, processHandler)
+    }
+
+    override fun createQuery(): Query {
+        val configuration = environment.runProfile as XRayTestConfiguration
+
+        val query = createQuery(environment.executor, configuration)
+        query.updating = false
+        query.database = configuration.database ?: ""
+        query.server = configuration.server ?: ""
+        query.modulePath = configuration.modulePath ?: ""
+
+        configuration.testPath?.let { query.bindVariable("test-dir", it, "xs:string") }
+        configuration.modulePattern?.let { query.bindVariable("module-pattern", it, "xs:string") }
+        configuration.testPattern?.let { query.bindVariable("test-pattern", it, "xs:string") }
+        query.bindVariable("format", configuration.outputFormat.name.toLowerCase(), "xs:string")
+        return query
+    }
+
+    private fun createQuery(executor: Executor, configuration: XRayTestConfiguration): Query {
+        val source = XRayTestService.getInstance(environment.project).runTestsQuery!!
+        val session = configuration.processor!!.session
+        return when (executor.id) {
+            DefaultRunExecutor.EXECUTOR_ID -> {
+                (session as RunnableQueryProvider).createRunnableQuery(source, XQuery)
+            }
+            else -> throw UnsupportedOperationException()
+        }
+    }
+
+    override fun createProcessHandler(query: Query): ProcessHandler = when (query) {
+        is RunnableQuery -> RunnableQueryProcessHandler(query)
+        else -> throw UnsupportedOperationException()
+    }
+
+    override fun createConsole(executor: Executor): ConsoleView = when (executor.id) {
+        DefaultRunExecutor.EXECUTOR_ID, DefaultDebugExecutor.EXECUTOR_ID -> {
+            QueryConsoleView(environment.project, QueryTextConsoleView(environment.project))
+        }
+        else -> throw UnsupportedOperationException()
     }
 }